+++ archives/MemoryCounter.h
@@ -194,7 +194,7 @@ namespace Util
    template <typename T, size_t N>
    inline MemoryCounter& MemoryCounter::operator << (T (& data)[N])
    {   
-      for (int i = 0; i < N; ++i) {
+      for (size_t i = 0; i < N; ++i) {
          serialize(*this, data[i], version_); 
       }
       return *this;
@@ -206,7 +206,7 @@ namespace Util
    template <typename T, size_t N>
    inline MemoryCounter& MemoryCounter::operator & (T (& data)[N])
    {   
-      for (int i = 0; i < N; ++i) {
+      for (size_t i = 0; i < N; ++i) {
          serialize(*this, data[i], version_); 
       }
       return *this;
