+++ archives/MemoryIArchive.h
@@ -283,7 +283,7 @@ namespace Util
    template <typename T, size_t N>
    inline MemoryIArchive& MemoryIArchive::operator >> (T (&data)[N])
    {
-      for (int i=0; i < N; ++i) {   
+      for (size_t i=0; i < N; ++i) {   
          serialize(*this, data[i], version_); 
       }
       return *this;
@@ -295,7 +295,7 @@ namespace Util
    template <typename T, size_t N>
    inline MemoryIArchive& MemoryIArchive::operator & (T (&data)[N])
    {
-      for (int i=0; i < N; ++i) {
+      for (size_t i=0; i < N; ++i) {
          serialize(*this, data[i], version_); 
       }
       return *this;
